requirements.txt
black==22.6.0
click==8.1.3
mypy-extensions==0.4.3
mysql-connector-python==8.0.29
pathspec==0.9.0
platformdirs==2.5.2
protobuf==4.21.2
PyYAML==6.0
tomli==2.0.1
typing-extensions==4.3.0
